requirements.txt
falcon
gunicorn
git+https://github.com/y-du/simple-env-var-manager.git@1.0.2
paho-mqtt
